Add files via upload
This program is designed to take user input on 5 different events, their cost, and the total number of people attending and outputs the events by sorting them out from lowest cost to highest. As the maximum number of people allowed are 100, if the user inputs more than a hundred for any of the events, it will post it on top as well as remove the event automatically.